Unify binaries into single spicebench binary with subcommands
