+++ .github/workflows/pr.yml
@@ -121,19 +121,13 @@ jobs:
 
       - uses: ./.github/actions/build-spicebench
 
-      - name: Build data-generation
-        run: |
-          mkdir -p ~/.spice/bin
-          cargo build -p data-generation
-          install -m 755 target/debug/data-generation ~/.spice/bin/data-generation
-
       - name: Generate test data (sf 0.001)
         env:
           AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
           AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
           RUST_LOG: info
         run: |
-          ~/.spice/bin/data-generation run \
+          ~/.spice/bin/spicebench generate \
             --dataset tpch \
             --scenario tpch \
             --scale-factor 0.001 \
@@ -158,7 +152,7 @@ jobs:
           SPIDAPTER_ICEBERG_CATALOG_FROM: iceberg:https://glue.us-west-1.amazonaws.com/iceberg/v1/catalogs/211125479522/namespaces
           RUST_LOG: info
         run: |
-          ~/.spice/bin/spicebench \
+          ~/.spice/bin/spicebench run \
             --concurrency "${NUM_QUERY_CLIENTS}" \
             --scenario tpch \
             --etl-bucket spiceai-public-datasets \
