+++ Cargo.toml
@@ -150,6 +150,7 @@ databricks = []
 debezium = []
 delta_lake = []
 dremio = []
+duckdb = ["etl/duckdb"]
 flightsql = []
 ftp = []
 imap = []
